Change `When` property type from `long` to `int`
The `When` property in the `ChatMessageCommand` class was updated from `long` to `int`. This change affects the range of values the property can store.

+++ L4D2PlayStats.Web/L4D2PlayStats/GameInfo/Commands/ChatMessageCommand.cs
@@ -4,7 +4,7 @@ namespace L4D2PlayStats.GameInfo.Commands;
 
 public class ChatMessageCommand
 {
-    public long When { get; set; }
+    public int When { get; set; }
     public bool Public { get; set; }
     public Team Team { get; set; }
     public string? CommunityId { get; set; }
